[INFRA] Simplify the configuration of the TypeScript Rollup project
Move the whole code to the `index.ts` file. We previously add an extra indirection to an `app/initialize.ts` file that
didn't help to demonstrate the integration.

Simply the commands to run the dev server: only keep the watch mode to make experiments easier. The changes are always
reflected in the browser in development.

Simplify the rollup configuration
  - remove the unnecessary json plugin
  - only keep a DEV and PRODUCTION mode

Remove eslint and prettier dependencies and configuration. They added noise as they are not related to the bpmn-visualization
integration in a Rollup project.

Bump all devDependencies
