--- examples/projects/typescript-vanilla-with-rollup/.eslintignore
@@ -1,3 +1,0 @@
-.idea/*
-dist/*
-node_modules/*
