--- examples/projects/typescript-vanilla-with-rollup/.eslintrc.js
@@ -1,31 +1,0 @@
-module.exports = {
-  plugins: ["notice"],
-  parser: "@typescript-eslint/parser", // Specifies the ESLint parser
-  extends: [
-    "plugin:@typescript-eslint/recommended", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
-    "plugin:prettier/recommended", // Enables eslint-plugin-prettier and eslint-config-prettier. This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
-  ],
-  parserOptions: {
-    ecmaVersion: 2018, // Allows for the parsing of modern ECMAScript features
-    sourceType: "module", // Allows for the use of imports
-  },
-  rules: {
-    // Place to specify ESLint rules. Can be used to overwrite rules specified from the extended configs
-    "@typescript-eslint/explicit-function-return-type": [
-      "warn",
-      {
-        allowExpressions: true,
-        allowTypedFunctionExpressions: true,
-      },
-    ],
-    "no-console": ["error", { allow: ["warn", "error"] }],
-  },
-  overrides: [
-    {
-      files: ["NOTICE"],
-      rules: {
-        "@typescript-eslint/explicit-function-return-type": "off",
-      },
-    },
-  ],
-};
